+++ CHANGELOG.md
@@ -20,6 +20,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed --->
 
+## 0.9.0 - 2021-11-02
+### Added
+- Model audit logging
+- Config for spatie permissions
+### Fixed
+- Address types teams support
+
 ## 0.8.1 - 2021-10-28
 ### Fixed
 - Editing roles in teams mode